Added x86_64 SHA-NI acceleration for SHA224 and SHA256, refactored SHA2
